+++ exporter/exporterhelper/logs_test.go
@@ -55,7 +55,7 @@ func TestLogsRequest(t *testing.T) {
 	assert.EqualValues(
 		t,
 		newLogsRequest(context.Background(), plog.NewLogs(), nil),
-		lr.onError(logErr),
+		lr.OnError(logErr),
 	)
 }