+++ pdata/internal/metrics.go
@@ -183,7 +183,31 @@ func (at MetricAggregationTemporality) String() string {
 	return otlpmetrics.AggregationTemporality(at).String()
 }
 
-// MetricDataPointFlagsStruct defines how a metric aggregator reports aggregated values.
+// FlagsStruct returns the flagsstruct associated with this NumberDataPoint.
+// Deprecated [0.58.0] Use Flags() instead
+func (ms NumberDataPoint) FlagsStruct() MetricDataPointFlags {
+	return ms.Flags()
+}
+
+// FlagsStruct returns the flagsstruct associated with this HistogramDataPoint.
+// Deprecated [0.58.0] Use Flags() instead
+func (ms HistogramDataPoint) FlagsStruct() MetricDataPointFlags {
+	return ms.Flags()
+}
+
+// FlagsStruct returns the flagsstruct associated with this ExponentialHistogramDataPoint.
+// Deprecated [0.58.0] Use Flags() instead
+func (ms ExponentialHistogramDataPoint) FlagsStruct() MetricDataPointFlags {
+	return ms.Flags()
+}
+
+// FlagsStruct returns the flagsstruct associated with this SummaryDataPoint.
+// Deprecated [0.58.0] Use Flags() instead
+func (ms SummaryDataPoint) FlagsStruct() MetricDataPointFlags {
+	return ms.Flags()
+}
+
+// MetricDataPointFlags defines how a metric aggregator reports aggregated values.
 // It describes how those values relate to the time interval over which they are aggregated.
 //
 // This is a reference type, if passed by value and callee modifies it the
@@ -191,42 +215,42 @@ func (at MetricAggregationTemporality) String() string {
 //
 // Must use NewMetricDataPointFlagsStruct function to create new instances.
 // Important: zero-initialized instance is not valid for use.
-type MetricDataPointFlagsStruct struct {
+type MetricDataPointFlags struct {
 	orig *uint32
 }
 
-func newMetricDataPointFlagsStruct(orig *uint32) MetricDataPointFlagsStruct {
-	return MetricDataPointFlagsStruct{orig: orig}
+func newMetricDataPointFlags(orig *uint32) MetricDataPointFlags {
+	return MetricDataPointFlags{orig: orig}
 }
 
-// NewMetricDataPointFlagsStruct creates a new empty MetricDataPointFlagsStruct.
+// NewMetricDataPointFlags creates a new empty MetricDataPointFlags.
 //
 // This must be used only in testing code. Users should use "AppendEmpty" when part of a Slice,
 // OR directly access the member if this is embedded in another struct.
-func NewMetricDataPointFlagsStruct() MetricDataPointFlagsStruct {
-	return newMetricDataPointFlagsStruct(new(uint32))
+func NewMetricDataPointFlags() MetricDataPointFlags {
+	return newMetricDataPointFlags(new(uint32))
 }
 
 // MoveTo moves all properties from the current struct to dest
 // resetting the current instance to its zero value
-func (ms MetricDataPointFlagsStruct) MoveTo(dest MetricDataPointFlagsStruct) {
+func (ms MetricDataPointFlags) MoveTo(dest MetricDataPointFlags) {
 	*dest.orig = *ms.orig
 	*ms.orig = uint32(otlpmetrics.DataPointFlags_FLAG_NONE)
 }
 
 // CopyTo copies all properties from the current struct to the dest.
-func (ms MetricDataPointFlagsStruct) CopyTo(dest MetricDataPointFlagsStruct) {
+func (ms MetricDataPointFlags) CopyTo(dest MetricDataPointFlags) {
 	*dest.orig = *ms.orig
 }
 
 // NoRecordedValue returns true if the MetricDataPointFlags contains the NO_RECORDED_VALUE flag.
-func (ms MetricDataPointFlagsStruct) NoRecordedValue() bool {
+func (ms MetricDataPointFlags) NoRecordedValue() bool {
 	return *ms.orig&uint32(otlpmetrics.DataPointFlags_FLAG_NO_RECORDED_VALUE) != 0
 }
 
 // SetNoRecordedValue sets the FLAG_NO_RECORDED_VALUE flag if true and removes it if false.
 // Setting this Flag when it is already set will change nothing.
-func (ms MetricDataPointFlagsStruct) SetNoRecordedValue(b bool) {
+func (ms MetricDataPointFlags) SetNoRecordedValue(b bool) {
 	if b {
 		*ms.orig |= uint32(otlpmetrics.DataPointFlags_FLAG_NO_RECORDED_VALUE)
 	} else {
@@ -235,7 +259,7 @@ func (ms MetricDataPointFlagsStruct) SetNoRecordedValue(b bool) {
 }
 
 // String returns the string representation of the MetricDataPointFlags.
-func (ms MetricDataPointFlagsStruct) String() string {
+func (ms MetricDataPointFlags) String() string {
 	return otlpmetrics.DataPointFlags(*ms.orig).String()
 }