+++ pdata/pmetric/alias.go
@@ -52,10 +52,19 @@ const (
 
 // MetricDataPointFlagsStruct defines how a metric aggregator reports aggregated values.
 // It describes how those values relate to the time interval over which they are aggregated.
-type MetricDataPointFlagsStruct = internal.MetricDataPointFlagsStruct
+// Deprecated [0.58.0] Use MetricDataPointFlags instead
+type MetricDataPointFlagsStruct = MetricDataPointFlags
 
 // NewMetricDataPointFlagsStruct returns a new empty MetricDataPointFlagsStruct.
-var NewMetricDataPointFlagsStruct = internal.NewMetricDataPointFlagsStruct
+// Deprecated [0.58.0] Use NewMetricDataPointFlags instead
+var NewMetricDataPointFlagsStruct = NewMetricDataPointFlags
+
+// MetricDataPointFlags defines how a metric aggregator reports aggregated values.
+// It describes how those values relate to the time interval over which they are aggregated.
+type MetricDataPointFlags = internal.MetricDataPointFlags
+
+// NewMetricDataPointFlags returns a new empty MetricDataPointFlags.
+var NewMetricDataPointFlags = internal.NewMetricDataPointFlags
 
 // NumberDataPointValueType specifies the type of NumberDataPoint value.
 type NumberDataPointValueType = internal.NumberDataPointValueType