+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 ### 🚩 Deprecations 🚩
 
 - Deprecate duplicate settings in service.ConfigProvider, embed ResolverSettings (#5843)
-- Deprecate `featuregate.Registry.MustApply` in favor of `featuregate.Registry.Apply`.
+- Deprecate `featuregate.Registry.MustApply` in favor of `featuregate.Registry.Apply`. (#5801)
 - Deprecate the `component.Factory.StabilityLevel(config.DataType)` in favor of Stability per component (#5762):
   - `component.ExporterFactory.TracesExporterStability`
   - `component.ExporterFactory.MetricsExporterStability`
@@ -37,12 +37,17 @@
   - `component.ReceiverFactory.MetricsReceiverStability`
   - `component.ReceiverFactory.LogsReceiverStability`
 - Deprecate `obsreport.ProcessorSettings.Level` and `obsreport.ExporterSettings.Level`, use MetricsLevel from CreateSettings (#5824)
+- Deprecates `FlagsStruct` in favor of `Flags` (#5842)
+  - `MetricDataPointFlagsStruct` -> `MetricDataPointFlags`
+  - `NewMetricDataPointFlagsStruct` -> `NewMetricDataPointFlags`
+  - `FlagsStruct` -> `Flags`
 
 ### 💡 Enhancements 💡
 
 - Enable persistent queue in the build by default (#5828)
 - Bump to opentelemetry-proto v0.19.0. (#5823)
 - Expose `Scope.Attributes` in pdata (#5826)
+- Add support to handle 404, 405 http error code as permanent errors in OTLP exporter (#5827)
 
 ### 🧰 Bug fixes 🧰